requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
joblib>=1.2.0
pyyaml>=6.0
scipy>=1.10.0
torch>=2.0.0
transformers>=4.30.0
underthesea>=6.3.0
